requirements.txt
argon2-cffi==20.1.0
cffi==1.14.0
dj-database-url==0.5.0
Django==2.1.2
django-betterforms==1.2
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
html5lib==1.0.1
passlib==1.7.2
pdfkit==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyPDF2==1.26.0
python-decouple==3.3
pytz==2019.3
reportlab==3.5.42
six==1.15.0
webencodings==0.5.1
whitenoise==5.1.0
xhtml2pdf==0.2.4
